+++ src/components/molecules/index.js
@@ -10,7 +10,7 @@ import UserDetail from './UserDetail/index.vue';
 import UserList from './UserList/index.vue';
 import CategoryPost from './CategoryPost/index.vue';
 import CategoryList from './CategoryList/index.vue';
-import CategoryDetail from './CategoryDetail/index.vue';
+import CategoryEdit from './CategoryEdit/index.vue';
 import ArticleEdit from './ArticleEdit/index.vue';
 import ArticlePost from './ArticlePost/index.vue';
 import ArticleDetail from './ArticleDetail/index.vue';
@@ -30,7 +30,7 @@ export {
   UserList,
   CategoryPost,
   CategoryList,
-  CategoryDetail,
+  CategoryEdit,
   ArticleEdit,
   ArticlePost,
   ArticleDetail,
